fix: 代码审计问题修复
- 切片前检查空字符串 (6处)
- interest_boost 返回类型改为 float
- 提取 bucket_data 初始化为私有方法

+++ config.py
@@ -228,7 +228,7 @@ class PluginConfig:
 
     @property
     def interest_boost(self):
-        return int(self._config.get("interest_boost", 2))
+        return float(self._config.get("interest_boost", 2.0))
 
     @property
     def daily_chat_boost(self):
